expose VarLenFeatureWithRank as public API
