+++ tensorflow_io/core/python/api/experimental/columnar.py
@@ -29,3 +29,8 @@ from tensorflow_io.core.python.experimental.parse_avro_ops import (  # pylint: d
 from tensorflow_io.core.python.experimental.avro_dataset_ops import (  # pylint: disable=unused-import
     make_avro_dataset,
 )
+
+from tensorflow_io.core.python.experimental.varlen_feature_with_rank import (  # pylint:
+    # disable=unused-import
+    VarLenFeatureWithRank,
+)
