+++ tests/test_parse_avro_eager.py
@@ -26,11 +26,10 @@ import tensorflow as tf
 from avro.io import DatumReader, DatumWriter, BinaryDecoder, BinaryEncoder
 from avro.datafile import DataFileReader, DataFileWriter
 from avro.schema import Parse as parse
-from tensorflow_io.core.python.experimental.parse_avro_ops import VarLenFeatureWithRank
 import tensorflow_io as tfio
 
-# if sys.platform == "darwin":
-#    pytest.skip("TODO: skip macOS", allow_module_level=True)
+if sys.platform == "darwin":
+   pytest.skip("TODO: skip macOS", allow_module_level=True)
 
 
 class AvroRecordsToFile:
@@ -427,7 +426,7 @@ class MakeAvroRecordDatasetTest(AvroDatasetTestBase):
             {"int_list_list": [[6]]},
             {"int_list_list": [[6]]},
         ]
-        features = {"int_list_list[*][*]": VarLenFeatureWithRank(tf.dtypes.int32)}
+        features = {"int_list_list[*][*]": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.int32)}
         expected_data = [
             {
                 "int_list_list[*][*]": tf.compat.v1.SparseTensorValue(
@@ -481,7 +480,7 @@ class MakeAvroRecordDatasetTest(AvroDatasetTestBase):
             {"int_list_list": [[6]]},
             {"int_list_list": [[6]]},
         ]
-        features = {"int_list_list[*][*]": VarLenFeatureWithRank(tf.dtypes.int32, 2)}
+        features = {"int_list_list[*][*]": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.int32, 2)}
         expected_data = [
             {
                 "int_list_list[*][*]": tf.compat.v1.SparseTensorValue(
@@ -1369,7 +1368,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
                   }
               ]}"""
         record_data = [{"int_list": [1, 2]}, {"int_list": [3, 4, 5]}, {"int_list": [6]}]
-        features = {"int_list[*]": VarLenFeatureWithRank(tf.dtypes.int32, 1)}
+        features = {"int_list[*]": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.int32, 1)}
         expected_data = [
             {
                 "int_list[*]": tf.compat.v1.SparseTensorValue(
@@ -1410,7 +1409,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
             {"int_list_list": [[6]]},
             {"int_list_list": [[6]]},
         ]
-        features = {"int_list_list[*][*]": VarLenFeatureWithRank(tf.dtypes.int32, 2)}
+        features = {"int_list_list[*][*]": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.int32, 2)}
         expected_data = [
             {
                 "int_list_list[*][*]": tf.compat.v1.SparseTensorValue(
@@ -1781,8 +1780,8 @@ class AvroDatasetTest(AvroDatasetTestBase):
             },
         ]
         features = {
-            "guests[gender='male'].name": VarLenFeatureWithRank(tf.dtypes.string),
-            "guests[gender='female'].name": VarLenFeatureWithRank(tf.dtypes.string),
+            "guests[gender='male'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string),
+            "guests[gender='female'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string),
         }
         expected_data = [
             {
@@ -1850,7 +1849,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
             {"guests": [{"name": "Joel", "gender": "male"}]},
         ]
         features = {
-            "guests[gender='wrong_value'].name": VarLenFeatureWithRank(tf.dtypes.string)
+            "guests[gender='wrong_value'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string)
         }
         expected_data = [
             {
@@ -1898,7 +1897,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
           """
         record_data = [{"guests": [{"name": "Hans"}]}]
         features = {
-            "guests[wrong_key='female'].name": VarLenFeatureWithRank(tf.dtypes.string)
+            "guests[wrong_key='female'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string)
         }
         self._test_fail_dataset(reader_schema, record_data, features, 1)
 
@@ -1931,7 +1930,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
           """
         record_data = [{"guests": [{"name": "Hans"}]}]
         features = {
-            "guests[forgot_the_separator].name": VarLenFeatureWithRank(tf.dtypes.string)
+            "guests[forgot_the_separator].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string)
         }
         self._test_fail_dataset(reader_schema, record_data, features, 1)
 
@@ -1964,7 +1963,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
           """
         record_data = [{"guests": [{"name": "Hans"}]}]
         features = {
-            "guests[used=too=many=separators].name": VarLenFeatureWithRank(
+            "guests[used=too=many=separators].name": tfio.experimental.columnar.VarLenFeatureWithRank(
                 tf.dtypes.string
             )
         }
@@ -2043,7 +2042,7 @@ class AvroDatasetTest(AvroDatasetTestBase):
             }
         ]
         features = {
-            "guests[gender='female'].address.street": VarLenFeatureWithRank(
+            "guests[gender='female'].address.street": tfio.experimental.columnar.VarLenFeatureWithRank(
                 tf.dtypes.string
             )
         }
@@ -2113,8 +2112,8 @@ class AvroDatasetTest(AvroDatasetTestBase):
             },
         ]
         features = {
-            "guests[gender='male'].name": VarLenFeatureWithRank(tf.dtypes.string),
-            "guests[gender='female'].name": VarLenFeatureWithRank(tf.dtypes.string),
+            "guests[gender='male'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string),
+            "guests[gender='female'].name": tfio.experimental.columnar.VarLenFeatureWithRank(tf.dtypes.string),
         }
         expected_data = [
             {
